fix failures due to contraction in ssui, 5.9
